3.0 beta版本 1.总结框的支持多行显示 2.多行显示闪退的bug修复 3.总结框多行显示有概率出现重复的bug修复

+++ app/src/main/java/com/example/testshareperference/activity/entity/Weather.java
@@ -2,14 +2,14 @@ package com.example.testshareperference.activity.entity;
 
 public class Weather {
 
-    private int code;
+    private double code;
     private String msg;
     private WeatherData data;
 
-    public void setCode(int code) {
+    public void setCode(double code) {
         this.code = code;
     }
-    public int getCode() {
+    public double getCode() {
         return code;
     }
